requirements.txt
Django==4.2.4
celery==5.3.1
gunicorn==21.2.0
uvicorn==0.23.2
djangorestframework==3.14.0
black==23.7.0
